Improve BattleChit border code.  Still not perfect. Change BrownColossus color to be harder to confuse with black. Add test main() to PickColor Remove exception in PredictSplits.merge()  Just print and abort.
git-svn-id: svn://localhost/tmp/s2g/colossus.svn/trunk@2093 c5679a2a-0e0e-0410-bba2-8642b794b8c1

+++ Colossus/net/sf/colossus/client/Client.java
@@ -1236,7 +1236,7 @@ public final class Client implements IClient
         {
             colorName = getColorByMarkerId(attackerMarkerId);
         }
-        BattleChit chit = new BattleChit(5 * Scale.get(), imageName,
+        BattleChit chit = new BattleChit((int)(5 * Scale.get()), imageName,
             map, inverted, tag, hexLabel, colorName);
         battleChits.add(chit);
     }
